# Conflicts:
#	backend/ziip/src/main/java/com/lastdance/ziip/global/exception/validator/MemberValidator.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/controller/MemberController.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/dto/LoginDto.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/dto/TokenDto.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/dto/request/MemberInfoUpdateRequestDto.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/dto/request/NicknameRequestDto.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/dto/response/LoginResponseDto.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/dto/response/MemberInfoResponseDto.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/dto/response/MemberUpdateResponseDto.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/dto/response/NickNameResponseDto.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/dto/response/RefreshTokenResponseDto.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/repository/MemberRepository.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/service/MemberService.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/service/MemberServiceImpl.java
#	backend/ziip/src/main/resources/application.yml


Former-commit-id: b6d39b1a044b1f9bd7f094ae36807f3ada68dbb8

+++ backend/ziip/src/main/java/com/lastdance/ziip/member/dto/TokenDto.java
@@ -10,10 +10,10 @@ public class TokenDto {
 
     private String accessToken;
     private String refreshToken;
-    private Integer id;
+    private Long id;
 
     @Builder
-    public TokenDto(String accessToken, String refreshToken, Integer id) {
+    public TokenDto(String accessToken, String refreshToken, Long id) {
         this.accessToken = accessToken;
         this.refreshToken = refreshToken;
         this.id = id;
